+++ vzmi/ychaos/core/verification/plugins/BaseVerificationPlugin.py
@@ -2,6 +2,7 @@
 #  Licensed under the terms of the ${MY_OSI} license. See the LICENSE file in the project root for terms
 from abc import abstractmethod
 
+from vzmi.ychaos.app_logger import AppLogger
 from vzmi.ychaos.core.verification.data import (
     VerificationData,
     VerificationStateData,
@@ -20,6 +21,9 @@ class BaseVerificationPlugin:
     ):
         self.config = config
         self.state_data = state_data
+        self.logger = AppLogger.get_logger(self.__class__.__name__)
+
+        self.logger.bind(event="verification", type=self.__verification_type__)
 
     @abstractmethod
     def run_verification(self) -> VerificationStateData: